Rename certain files to match class names
