+++ hamsa_lomi/lib/presentation/lang_acct_type/pages/choose_account_type.dart
@@ -1,11 +1,11 @@
 // Flutter imports:
 import 'package:flutter/material.dart';
+import '../../core/hamsa_app_bar.dart';
 
 // Project imports:
 import '../../constants/app_assets_constant.dart';
 import '../../constants/app_string_constant.dart';
 import '../../create_account/pages/create_account_page.dart';
-import '../../share_widgets/appbar_widget.dart';
 import '../../theme/hamsa_theme.dart';
 import '../widget/button_widget.dart';
