+++ hamsa_lomi/lib/presentation/sign_in/pages/sign_in_page.dart
@@ -1,9 +1,9 @@
 // Flutter imports:
 import 'package:flutter/material.dart';
+import '../../core/hamsa_app_bar.dart';
+import '../../core/tabsbar_widget.dart';
 
 // Project imports:
-import '../../share_widgets/appbar_widget.dart';
-import '../../share_widgets/tabsbar_widget.dart';
 import '../widgets/sign_in_with_email_form.dart';
 import '../widgets/sign_in_with_phone_form.dart';
